refactor TaskDetails by extracting column related components
