+++ src/components/TaskDetails.tsx
@@ -1,6 +1,4 @@
 import { useState, createRef, useEffect } from 'react'
-import ChevronDown from './icons/ChevronDown'
-import ChevronUp from './icons/ChevronUp'
 import VerticalEllipsis from './icons/VerticalEllipsis'
 import ColumnSelector from './ColumnSelector'
 import EditMenu from './EditMenu'
@@ -17,11 +15,8 @@ export default function TaskDetails({
   complete: number
 }) {
   const [menuActive, setMenuActive] = useState<boolean>(false)
-  const [selectColumnActive, setSelectColumnActive] = useState<boolean>(false)
   const editMenuRef = createRef<HTMLDivElement>() // <- this
-  const columnSelectRef = createRef<HTMLDivElement>()
 
-  // Add event handler to detect if user clicks outside of edit modal
   useEffect(() => {
     if (!menuActive) return
     function handleEditMenuOutsideClick(e: MouseEvent) {
@@ -40,26 +35,6 @@ export default function TaskDetails({
     return () =>
       document.removeEventListener('click', handleEditMenuOutsideClick)
   }, [menuActive, editMenuRef])
-
-  // Same for column status selector
-  useEffect(() => {
-    if (!selectColumnActive) return
-    function handleColumnSelectOutsideClick(e: MouseEvent) {
-      if (
-        columnSelectRef.current &&
-        !columnSelectRef.current.contains(e.target as HTMLElement)
-      )
-        setSelectColumnActive((prev) => !prev)
-    }
-
-    setTimeout(
-      () => document.addEventListener('click', handleColumnSelectOutsideClick),
-      500
-    )
-
-    return () =>
-      document.removeEventListener('click', handleColumnSelectOutsideClick)
-  }, [selectColumnActive, columnSelectRef])
 
   return (
     <div className='task-details p-32px bg-white rounded-sm'>
@@ -86,23 +61,7 @@ export default function TaskDetails({
         </ul>
       </div>
 
-      <div className='column-selector-wrapper'>
-        <h3 className='column-selector-title text-med-gray text-xs font-bold mb-[0.8rem]'>
-          Current Status
-        </h3>
-        <div className='relative'>
-          <button
-            className='flex items-center justify-between w-full px-16px py-[0.8rem] border border-cool-gray rounded-sm'
-            onClick={() => setSelectColumnActive((prev) => !prev)}
-          >
-            <span className='text-sm font-medium leading-extra-loose'>
-              {'Doing'}
-            </span>
-            {selectColumnActive ? <ChevronUp /> : <ChevronDown />}
-          </button>
-          {selectColumnActive ? <ColumnSelector ref={columnSelectRef} /> : null}
-        </div>
-      </div>
+      <ColumnSelector />
     </div>
   )
 }
